% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spokes.R
\name{calc_spokes}
\alias{calc_spokes}
\title{Calculate the 'spokes' of the spider plot}
\usage{
calc_spokes(num)
}
\arguments{
\item{num}{The number of spokes}
}
\value{
A data frame with num rows and columns x, y, xend, and yend representing the x-y
start and end coordinates of the spokes
}
\description{
Calculate the 'spokes' of the spider plot
}
\examples{
\dontrun{
calc_spokes(5)
}
}
