% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GCDKitNormScheme}
\alias{GCDKitNormScheme}
\title{GCDkit Normalization schemes}
\format{
A list with 39 items.
Each item contains a named vector representing
one normalization scheme (names of elements in order and values)
}
\source{
GCDkit package, <https://gcdkit.org>>
}
\usage{
GCDKitNormScheme
}
\description{
Various normalization factors for spider plots
}
\keyword{datasets}
