% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggSpider.R
\name{geom_line_continuous}
\alias{geom_line_continuous}
\title{Lines that connect accross missing values in spidergrams}
\usage{
geom_line_continuous(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\description{
This is an exact equivalent to geom_line(), with the exception
that the line is connected across missing values.
}
\seealso{
[geom_line()]
}
