% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggSpider.R
\name{geom_range}
\alias{geom_range}
\title{A field encompassing all the values}
\usage{
geom_range(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\description{
This generates a polygon that includes all the values, for each element
}
