% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggSpider.R
\name{get_norm}
\alias{get_norm}
\title{Interpret the normalization definition}
\usage{
get_norm(norm, normScheme = GCDKitNormScheme)
}
\arguments{
\item{norm}{Named vector or char. Either normalization values, as a named vector; or
the name of a normalization to look for.}

\item{normScheme}{A tibble containing normalization names in the first column, and
named vectors in the second.}
}
\value{
A named vector with the normalization factors, in order
}
\description{
If the main argument, norm, is already a named vector, use it without further
action. Otherwise if it is a char string, try to match it with one of the
named normalizations from the normScheme database (normally, imported from
GCDkit's spider.data)
}
