% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggSpider.R
\name{ggspiderplot}
\alias{ggspiderplot}
\title{A function to build the skeleton of a spidergram}
\usage{
ggspiderplot(
  df,
  norm,
  .norm = get_norm(norm),
  .df = spider_data(df, norm),
  ...
)
}
\arguments{
\item{df}{Data frame containing compositional data to plot}

\item{norm}{named vector or string containing either the the normalization
values, or the name of a scheme. Normally processed by get_norm}

\item{.norm}{Named vector with normalization info, if the user needs
to bypass the call to get_norm)}

\item{.df}{Data frame already formatted for spidergrams. Normally not used,
only supplied if the user needs to override the call to spider_data()}
}
\value{
a ggplot object, to which geometries must be added.
}
\description{
A spidergram is a special type of diagram, using "parallel coordinates". This
is frowned upon in data sciences but commonly used in geochemistry, for better or
for worse!
From a ggplot perspective, the core definition of the spidergram includes
(i) mapping of the X-axis to element names; (ii) mapping of the Y-axis to elements
normalized values; (iii) using a log scale for Y (although this can be overridden latter,
if needed). All of the other mappings (colors, symbols etc) are let to the geometries.

Plotting such a diagram requires specially prepared data, which is normally done using
function spider_data().
}
