% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_attribute}
\alias{has_attribute}
\title{has_attribute()}
\usage{
has_attribute(x, which)
}
\arguments{
\item{x}{the object to text}

\item{which}{the attribute}
}
\value{
Boolean.
}
\description{
Check whtther an object has a specific attribute
}
\details{
That's really a simple function, what?
}
\examples{
\dontrun{
if(interactive()){
 has_attribute(mtcars,"names") ##TRUE
 has_attribute(mtcars,"dim") ##FALSE
 }
}
}
