% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggSpider.R
\name{spider_data}
\alias{spider_data}
\title{Prepare data to plot a spidergram}
\usage{
spider_data(df, norm)
}
\arguments{
\item{df}{a data.frame that will be formatted for spidergrams}

\item{norm}{A normalization scheme, either as a named vector or the name of a scheme existing
in the database. It will be processed by get_norm().}
}
\value{
a tibble in long form, with one line for each sample/element combination. All
the other variables are retained, to allow further aes() mappings.
}
\description{
The data used in a spiderplot needs to be (i) arranged in the prescribed order
and (ii) normalized to the reference values. In addition, (iii) since we are
not using a proper coorinate systems (X has no meaning), we are using a so-called
"parallel coordinates" plot and as such we must reformat the data to be used in this
way. The resulting data will not be usable for general purposes (if only because we create
one line per element/sample combination!)
}
