% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{calc_web}
\alias{calc_web}
\title{Title}
\usage{
calc_web(mydf)
}
\arguments{
\item{mydf}{A data frame. The first column holds the names for the 'web' data.
The remaining columns are the names of the 'spokes'.}
}
\value{
A data.frame in 'long' format with columns 'group', 'x', and 'y' where group
 holds the names of the 'web' data groups, with 'x' and 'y' representing points on each
 spoke, with each group having number of rows = number of spokes plus one (the first one
 repeats at the end so that plot lines will connect).
}
\description{
Title
}
\examples{
df <- data.frame(nm = c("A", "B", "C"),
                 grp1 = c(1, 2, 3),
                 grp2 = c(4, 5, 6))
web <- calc_web(df)
}
