% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{spider_web}
\alias{spider_web}
\title{Make a spider web plot}
\usage{
spider_web(df, grp_col = "nm", spk_col = "spk", value_col = "value",
  spoke_color = "grey75", spoke_lty = 1, ref_lines_val = c(0.5, 0.75,
  1), ref_lines_color = rep("grey75", length(ref_lines_val)),
  ref_lines_type = rep(1, length(ref_lines_val)),
  ref_label_color = rep("grey50", length(ref_lines_val)),
  ref_lines_label_spoke = c(length(unique(df[, spk_col, drop = TRUE])),
  length(unique(df[, spk_col, drop = TRUE])), 0),
  diff_lty_by_name = "ref", diff_lty = 2, lwd = 1,
  palette = "Set2", show_legend = TRUE, leg_main_title = "Legend",
  leg_lty_title = "Linetype")
}
\arguments{
\item{df}{A data.frame representing groups and spokes for the plot (see examples)}

\item{grp_col}{The name of the column used for groups}

\item{spk_col}{The name of the column used for spokes}

\item{value_col}{The name of the column used for the values of each group along the spokes}

\item{spoke_color}{Single value for color of the spokes}

\item{spoke_lty}{Single integer value for the line type of the spokes}

\item{ref_lines_val}{A vector of reference line values to plot}

\item{ref_lines_color}{A vector of reference line colors}

\item{ref_lines_type}{vector of reference line types}

\item{ref_label_color}{A vector of reference label colors}

\item{ref_lines_label_spoke}{A vector of reference label locations (spoke number). A value of
zero means that the label will not be printed}

\item{diff_lty_by_name}{This is an optional string which will be used to \code{\link[=grep]{grep()}} the names
of the values given in column grp_col. If they match, the \code{diff_lty} value will be used for
line type for those lines}

\item{diff_lty}{Line type to use for \code{grp_col} values that \code{diff_lty_by_name} matches with \code{\link[=grep]{grep()}}}

\item{lwd}{Line width for the main coloured lines.}

\item{palette}{As defined in \code{\link[ggplot2:scale_color_brewer]{ggplot2::scale_color_brewer()}}}

\item{show_legend}{Show legend?}

\item{leg_main_title}{Main legend title}

\item{leg_lty_title}{Linetype legend title}
}
\description{
Make a spider web plot
}
\examples{
df <- data.frame(
  nm = c(
    "A", "B", "C",
    "A", "B", "C",
    "A", "B", "C",
    "A", "B", "C",
    "A", "B", "C"
  ),
  spk = c(
    "P1", "P1", "P1",
    "P2", "P2", "P2",
    "P3", "P3", "P3",
    "P4", "P4", "P4",
    "P5", "P5", "P5"
  ),
  value = c(
    .1, .2, .3,
    .4, .5, .6,
    .7, .8, .9,
    .10, .11, .12,
    .13, .14, .15
  )
)
spider_web(df,
  ref_lines_label_spoke = c(4, 3, 0),
  ref_lines_type = c(2, 3, 4),
  ref_lines_color = c("black", "royalblue", "salmon")
)
}
